IntegrateAdminPages
Now I have Integrate the Admin Pages into one page and can switch each other. And I have enable them to share the same WebsocketService. So when switch pages it will still be the same page. But the base page is still the admin dashboard. If you refresh the page it will go back to dashboard instead of other current page.